# nadiki_registrar/openapi/openapi.yaml
openapi: 3.0.0
info:
  description: API for registering and managing data center facilities
  title: Facility Registry API
  version: 1.0.0
servers:
- description: Registry API v1
  url: https://api.registry.example.com/v1
...
paths:
...
  /servers:
    get:
      description: Retrieve a list of all registered servers
      operationId: list_servers
      parameters:
      - description: Maximum number of servers to return
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 20
          type: integer
        style: form
      - description: Number of servers to skip
        explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          type: integer
        style: form
      - description: Filter servers by facility ID
        explode: true
        in: query
        name: facility_id
        required: false
        schema:
          type: string
        style: form
      - description: Filter servers by rack ID
        explode: true
        in: query
        name: rack_id
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listServers_200_response'
          description: List of servers
      summary: List all servers
      tags:
      - server-api.spec_other
      x-openapi-router-controller: nadiki_registrar.controllers.server_api_spec_other_controller
    post:
      description: Create a new server entry in the registry
      operationId: create_server
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerCreate'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
          description: Server successfully created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid input
      summary: Register a new server
      tags:
      - server-api.spec_other
      x-openapi-router-controller: nadiki_registrar.controllers.server_api_spec_other_controller
  /servers/{serverId}:
    delete:
      description: Remove a server from the registry
      operationId: delete_server
      parameters:
      - description: "Unique server identifier (format SERVER-[FACILITY_ID]-[RACK_ID]-[SERVER_ID])"
        explode: false
        in: path
        name: serverId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Server successfully deleted
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server not found
      summary: Delete server
      tags:
      - server-api.spec_other
      x-openapi-router-controller: nadiki_registrar.controllers.server_api_spec_other_controller
    get:
      description: Retrieve detailed information about a specific server
      operationId: get_server
      parameters:
      - description: "Unique server identifier (format SERVER-[FACILITY_ID]-[RACK_ID]-[SERVER_ID])"
        explode: false
        in: path
        name: serverId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
          description: Server details
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server not found
      summary: Get server details
      tags:
      - server-api.spec_other
      x-openapi-router-controller: nadiki_registrar.controllers.server_api_spec_other_controller
    put:
      description: Update all server information
      operationId: update_server
      parameters:
      - description: "Unique server identifier (format SERVER-[FACILITY_ID]-[RACK_ID]-[SERVER_ID])"
        explode: false
        in: path
        name: serverId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerUpdate'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerResponse'
          description: Server successfully updated
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid input
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server not found
      summary: Update server
      tags:
      - server-api.spec_other
      x-openapi-router-controller: nadiki_registrar.controllers.server_api_spec_other_controller
  /servers/{serverId}/query:
    post:
      description: Retrieve aggregated time series metrics for a server over a specified
        time period
      operationId: query_server_metrics
      parameters:
      - description: "Unique server identifier (format SERVER-[FACILITY_ID]-[RACK_ID]-[SERVER_ID])"
        explode: false
        in: path
        name: serverId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerMetricsQuery'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerMetricsQueryResponse'
          description: Aggregated metrics data
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid query parameters
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Server not found
      summary: Query server metrics
      tags:
      - server-api.spec_other
      x-openapi-router-controller: nadiki_registrar.controllers.server_api_spec_other_controller
components:
  schemas:
...
    Error:
      properties:
        code:
          description: Error code
          title: code
          type: string
        message:
          description: Error message
          title: message
          type: string
        details:
          description: Additional error details
          title: details
          type: object
      required:
      - code
      - message
      title: Error
      type: object
...
    CPU:
      example:
        vendor: vendor
        type: type
      properties:
        vendor:
          description: "CPU vendor (e.g., Intel, AMD)"
          title: vendor
          type: string
        type:
          description: CPU type identifier
          title: type
          type: string
      required:
      - type
      - vendor
      title: CPU
      type: object
    GPU:
      example:
        vendor: vendor
        type: type
      properties:
        vendor:
          description: "GPU vendor (e.g., Nvidia, AMD)"
          title: vendor
          type: string
        type:
          description: GPU type identifier
          title: type
          type: string
      required:
      - type
      - vendor
      title: GPU
      type: object
    FPGA:
      example:
        vendor: vendor
        type: type
      properties:
        vendor:
          description: "FPGA vendor (e.g., Intel, Xilinx)"
          title: vendor
          type: string
        type:
          description: FPGA type identifier
          title: type
          type: string
      required:
      - type
      - vendor
      title: FPGA
      type: object
    HardDisk:
      example:
        vendor: vendor
        type: NVMe
        capacity: 2.302136
      properties:
        vendor:
          description: "Disk vendor (e.g., Samsung, Intel)"
          title: vendor
          type: string
        capacity:
          description: Disk capacity in TB
          format: float
          title: capacity
          type: number
        type:
          description: Type of storage device
          enum:
          - NVMe
          - SSD
          - HDD
          - Other
          title: type
          type: string
      required:
      - capacity
      - type
      - vendor
      title: HardDisk
      type: object
    ServerCreate:
      example:
        number_of_psus: 1
        rated_power: 0.8008282
        installed_gpus:
        - vendor: vendor
          type: type
        - vendor: vendor
          type: type
        total_fpgas: 0
        hard_disks:
        - vendor: vendor
          type: NVMe
          capacity: 2.302136
        - vendor: vendor
          type: NVMe
          capacity: 2.302136
        total_cpu_sockets: 1
        total_gpus: 0
        total_installed_memory: 5
        cooling_type: air
        rack_id: rack_id
        installed_fpgas:
        - vendor: vendor
          type: type
        - vendor: vendor
          type: type
        product_passport: "{}"
        installed_cpus:
        - vendor: vendor
          type: type
        - vendor: vendor
          type: type
        facility_id: facility_id
        number_of_memory_units: 5
      properties:
        facility_id:
          description: ID of the facility where the server is located
          title: facility_id
          type: string
        rack_id:
          description: ID of the rack where the server is installed
          title: rack_id
          type: string
        rated_power:
          description: Rated power in kW
          format: float
          title: rated_power
          type: number
        total_cpu_sockets:
          default: 2
          description: Total number of CPU sockets
          minimum: 1
          title: total_cpu_sockets
          type: integer
        installed_cpus:
          description: Array of installed CPUs
          items:
            $ref: '#/components/schemas/CPU'
          title: installed_cpus
          type: array
        number_of_psus:
          default: 2
          description: Number of power supply units
          minimum: 1
          title: number_of_psus
          type: integer
        total_installed_memory:
          description: Total installed memory in GB
          title: total_installed_memory
          type: integer
        number_of_memory_units:
          description: Number of memory units installed
          title: number_of_memory_units
          type: integer
        hard_disks:
          description: Array of installed hard disks
          items:
            $ref: '#/components/schemas/HardDisk'
          title: hard_disks
          type: array
        total_gpus:
          default: 0
          description: Total number of GPUs installed
          minimum: 0
          title: total_gpus
          type: integer
        installed_gpus:
          description: Array of installed GPUs
          items:
            $ref: '#/components/schemas/GPU'
          title: installed_gpus
          type: array
        total_fpgas:
          default: 0
          description: Total number of FPGAs installed
          minimum: 0
          title: total_fpgas
          type: integer
        installed_fpgas:
          description: Array of installed FPGAs
          items:
            $ref: '#/components/schemas/FPGA'
          title: installed_fpgas
          type: array
        product_passport:
          description: LCA product passport data
          title: product_passport
          type: object
        cooling_type:
          default: air
          description: Type of cooling system used for this server
          enum:
          - direct-to-chip
          - immersion
          - back-door-liquid
          - back-door-fan
          - air
          title: cooling_type
          type: string
      required:
      - cooling_type
      - facility_id
      - rack_id
      title: ServerCreate
      type: object
    ServerUpdate:
      allOf:
      - $ref: '#/components/schemas/ServerCreate'
      - required:
        - cooling_type
        - facility_id
        - rack_id
        type: object
      title: ServerUpdate
    ServerTimeSeriesDataPoint:
      properties:
        name:
          description: Prometheus metric name with unit suffix
          pattern: /^(cpu_energy_consumption_joules|server_energy_consumption_joules)$/
          title: name
          type: string
        unit:
          description: Prometheus metric family
          enum:
          - Time
          - Temperature
          - Length
          - Bytes
          - Percent
          - Voltage
          - Electric current
          - Energy
          - Power
          - Mass
          title: unit
          type: string
        granularitySeconds:
          default: 30
          description: Time series data granularity in seconds
          title: granularitySeconds
          type: integer
        labels:
          $ref: '#/components/schemas/ServerTimeSeriesDataPoint_labels'
      required:
      - granularitySeconds
      - labels
      - name
      - unit
      title: ServerTimeSeriesDataPoint
      type: object
    ServerTimeSeriesConfig:
      properties:
        endpoint:
          description: Time series endpoint URL with port number
          example: https://timeseries.registry.example.com:9090
          title: endpoint
          type: string
        dataPoints:
          description: List of time series metrics available for streaming
          items:
            $ref: '#/components/schemas/ServerTimeSeriesDataPoint'
          title: dataPoints
          type: array
      required:
      - dataPoints
      - endpoint
      title: ServerTimeSeriesConfig
      type: object
    ServerResponse:
      allOf:
      - $ref: '#/components/schemas/ServerCreate'
      - $ref: '#/components/schemas/ServerResponse_allOf'
      title: ServerResponse
    ServerMetricsQuery:
      example:
        startTime: 2000-01-23T04:56:07.000+00:00
        aggregation: sum
        endTime: 2000-01-23T04:56:07.000+00:00
        metrics:
        - metrics
        - metrics
      properties:
        startTime:
          description: Start timestamp for the query period
          format: date-time
          title: startTime
          type: string
        endTime:
          description: End timestamp for the query period
          format: date-time
          title: endTime
          type: string
        aggregation:
          description: Aggregation function to apply over the time period
          enum:
          - sum
          - average
          title: aggregation
          type: string
        metrics:
          description: "Optional list of specific metrics to query. If not provided,\
            \ all available metrics will be returned."
          items:
            pattern: ^(cpu_energy_consumption_joules|server_energy_consumption_joules)$
            type: string
          title: metrics
          type: array
      required:
      - aggregation
      - endTime
      - startTime
      title: ServerMetricsQuery
      type: object
    ServerMetricsQueryResponse:
      example:
        rack_id: rack_id
        facility_id: facility_id
        startTime: 2000-01-23T04:56:07.000+00:00
        aggregation: sum
        endTime: 2000-01-23T04:56:07.000+00:00
        metrics:
        - unit: Time
          granularitySeconds: 6
          name: name
          value: 0.8008282
        - unit: Time
          granularitySeconds: 6
          name: name
          value: 0.8008282
        server_id: server_id
      properties:
        facility_id:
          description: Facility identifier
          title: facility_id
          type: string
        rack_id:
          description: Rack identifier
          title: rack_id
          type: string
        server_id:
          description: "Server identifier (format SERVER-[FACILITY_ID]-[RACK_ID]-[SERVER_ID])"
          title: server_id
          type: string
        startTime:
          description: Start timestamp of the query period
          format: date-time
          title: startTime
          type: string
        endTime:
          description: End timestamp of the query period
          format: date-time
          title: endTime
          type: string
        aggregation:
          description: Aggregation function that was applied
          enum:
          - sum
          - average
          title: aggregation
          type: string
        metrics:
          items:
            $ref: '#/components/schemas/ServerMetricsQueryResponse_metrics_inner'
          title: metrics
          type: array
      required:
      - aggregation
      - endTime
      - facility_id
      - metrics
      - rack_id
      - server_id
      - startTime
      title: ServerMetricsQueryResponse
      type: object
...
    listServers_200_response:
      example:
        total: 0
        offset: 1
        limit: 6
        items:
        - null
        - null
      properties:
        items:
          items:
            $ref: '#/components/schemas/ServerResponse'
          title: items
          type: array
        total:
          description: Total number of servers
          title: total
          type: integer
        limit:
          description: Number of servers per page
          title: limit
          type: integer
        offset:
          description: Current offset
          title: offset
          type: integer
      title: listServers_200_response
      type: object
...
    ServerTimeSeriesDataPoint_labels:
      properties:
        facility_id:
          description: Facility identifier
          title: facility_id
          type: string
        rack_id:
          description: Rack identifier
          title: rack_id
          type: string
        server_id:
          description: "Server identifier (format SERVER-[FACILITY_ID]-[RACK_ID]-[SERVER_ID])"
          title: server_id
          type: string
        country_code:
          description: ISO 3166-1 alpha-3 country code
          title: country_code
          type: string
      required:
      - country_code
      - facility_id
      - rack_id
      - server_id
      title: ServerTimeSeriesDataPoint_labels
      type: object
    ServerResponse_allOf:
      properties:
        id:
          description: "Unique server identifier (format SERVER-[FACILITY_ID]-[RACK_ID]-[SERVER_ID])"
          example: SERVER-FACILITY-NLD-001-RACK-001-001
          title: id
          type: string
        timeSeriesConfig:
          $ref: '#/components/schemas/ServerTimeSeriesConfig'
        createdAt:
          format: date-time
          title: createdAt
          type: string
        updatedAt:
          format: date-time
          title: updatedAt
          type: string
      required:
      - id
      - timeSeriesConfig
      title: ServerResponse_allOf
      type: object
      example: null
    ServerMetricsQueryResponse_metrics_inner:
      example:
        unit: Time
        granularitySeconds: 6
        name: name
        value: 0.8008282
      properties:
        name:
          description: Name of the metric
          pattern: ^(cpu_energy_consumption_joules|server_energy_consumption_joules)$
          title: name
          type: string
        value:
          description: Aggregated value over the time period
          format: float
          title: value
          type: number
        unit:
          description: Unit of the metric
          enum:
          - Time
          - Temperature
          - Length
          - Bytes
          - Percent
          - Voltage
          - Electric current
          - Energy
          - Power
          - Mass
          title: unit
          type: string
        granularitySeconds:
          description: Original granularity of the data points in seconds
          title: granularitySeconds
          type: integer
      required:
      - granularitySeconds
      - name
      - unit
      - value
      title: ServerMetricsQueryResponse_metrics_inner
      type: object
...